fix: 변수 타입 Long -> Int로 수정

+++ src/main/java/im/toduck/domain/concentration/common/mapper/ConcentrationMapper.java
@@ -20,7 +20,7 @@ public class ConcentrationMapper {
 			.targetCount(concentration.getTargetCount())
 			.settingCount(concentration.getSettingCount())
 			.time(concentration.getTime())
-			.percentage((long)percentage)
+			.percentage(percentage)
 			.build();
 	}
